+++ subsys/mgmt/mcumgr/lib/cmd/fs_mgmt/src/zephyr_fs_mgmt.c
@@ -7,7 +7,6 @@
 #include <zephyr/fs/fs.h>
 #include <errno.h>
 #include <mgmt/mgmt.h>
-#include <zephyr/mgmt/mcumgr/buf.h>
 #include <fs_mgmt/fs_mgmt_impl.h>
 
 int
