+++ subsys/mgmt/mcumgr/lib/cmd/img_mgmt/src/img_mgmt_state.c
@@ -10,14 +10,14 @@
 
 #include <string.h>
 #include "img_mgmt/img_mgmt.h"
-#include <zephyr/mgmt/mcumgr/buf.h>
 #include "img_mgmt/image.h"
 #include "img_mgmt_priv.h"
 #include "img_mgmt/img_mgmt_impl.h"
-#include <mgmt/mgmt.h>
 #include <zcbor_common.h>
 #include <zcbor_decode.h>
 #include <zcbor_encode.h>
+#include <mgmt/mgmt.h>
+#include "smp/smp.h"
 #include "zcbor_bulk/zcbor_bulk_priv.h"
 
 /* The value here sets how many "characteristics" that describe image is
