+++ subsys/mgmt/mcumgr/lib/smp/include/smp/smp.h
@@ -25,14 +25,41 @@
 #ifndef H_SMP_
 #define H_SMP_
 
-#include "mgmt/mgmt.h"
+#include <zephyr/net/buf.h>
+#include <zephyr/mgmt/mcumgr/smp.h>
+#include <zcbor_common.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-struct smp_streamer;
-struct mgmt_hdr;
+struct cbor_nb_reader {
+	struct net_buf *nb;
+	/* CONFIG_MGMT_MAX_DECODING_LEVELS + 2 translates to minimal
+	 * zcbor backup states.
+	 */
+	zcbor_state_t zs[CONFIG_MGMT_MAX_DECODING_LEVELS + 2];
+};
+
+struct cbor_nb_writer {
+	struct net_buf *nb;
+	zcbor_state_t zs[2];
+};
+
+/**
+ * @brief Allocates a net_buf for holding an mcumgr request or response.
+ *
+ * @return                      A newly-allocated buffer net_buf on success;
+ *                              NULL on failure.
+ */
+struct net_buf *mcumgr_buf_alloc(void);
+
+/**
+ * @brief Frees an mcumgr net_buf
+ *
+ * @param nb                    The net_buf to free.
+ */
+void mcumgr_buf_free(struct net_buf *nb);
 
 /**
  * @brief Decodes, encodes, and transmits SMP packets.
