+++ subsys/mgmt/mcumgr/transport/smp_dummy.c
@@ -22,9 +22,9 @@
 #include <zephyr/drivers/console/uart_mcumgr.h>
 #include <mgmt/mgmt.h>
 #include <zephyr/mgmt/mcumgr/serial.h>
-#include <zephyr/mgmt/mcumgr/buf.h>
 #include <zephyr/mgmt/mcumgr/smp.h>
 #include <zephyr/mgmt/mcumgr/smp_dummy.h>
+#include <smp/smp.h>
 #include "../smp_internal.h"
 
 BUILD_ASSERT(CONFIG_MCUMGR_SMP_DUMMY_RX_BUF_SIZE != 0,
