+++ subsys/mgmt/mcumgr/transport/smp_udp.c
@@ -19,8 +19,8 @@
 #endif
 #include <errno.h>
 #include <mgmt/mgmt.h>
+#include "smp/smp.h"
 #include <zephyr/mgmt/mcumgr/smp_udp.h>
-#include <zephyr/mgmt/mcumgr/buf.h>
 #include <zephyr/mgmt/mcumgr/smp.h>
 #include "../smp_internal.h"
